+++ MERN/Client/student/src/App.js
@@ -1,52 +1,12 @@
-import React, { useState } from "react";
+import React from 'react'
+import SignIn from './Signin'
+import SignUp from './SignUp'
 
 export default function App() {
-  const [username, setUsername] = useState("");
-  const [email, setEmail] = useState("");
-  const [password, setPassword] = useState("");
-  const handleSubmit = async (e) => {
-    e.preventDefault();
-    const response =await fetch("http://localhost:3000/register", {
-      method: "POST",
-      headers: { "content-type": "application/json" },
-      body: JSON.stringify({ username, email, password }),
-    });
-  };
   return (
     <div>
-      <form onSubmit={handleSubmit}>
-        <h1>Sign Up</h1>
-        <div>
-          <label>Username</label>
-          <input
-            type="text"
-            name="username"
-            placeholder="Enter Username"
-            value={username}
-            onChange={(e) => setUsername(e.target.value)}
-          />
-        </div>
-        <div>
-          <label>Email</label>
-          <input
-            type="email"
-            name="email"
-            placeholder="Enter Email"
-            value={email}
-            onChange={(e) => setEmail(e.target.value)}
-          />
-        </div>
-        <div>
-          <label>Password</label>
-          <input
-            type="password"
-            name="password"
-            placeholder="Enter Password"
-            value={password}
-            onChange={(e) => setPassword(e.target.value)}
-          />
-        </div>
-      </form>
+      <SignUp />
+      <SignIn />
     </div>
-  );
+  )
 }
