+++ MERN/Client/student/src/SignUp.js
@@ -1,0 +1,58 @@
+import React, { useState } from "react";
+
+export default function SignUp() {
+  const [username, setUsername] = useState("");
+  const [email, setEmail] = useState("");
+  const [password, setPassword] = useState("");
+  const handleSubmit = async (e) => {
+    e.preventDefault();
+    const response =await fetch("http://localhost:3000/register", {
+      method: "POST",
+      headers: { "content-type": "application/json" },
+      body: JSON.stringify({ username, email, password }),
+    });
+    const data = await response.json();
+    console.log(data);
+    setUsername("");
+    setEmail("");
+    setPassword("");
+  };
+  return (
+    <div>
+      <form onSubmit={handleSubmit}>
+        <h1>Sign Up</h1>
+        <div>
+          <label>Username</label>
+          <input
+            type="text"
+            name="username"
+            placeholder="Enter Username"
+            value={username}
+            onChange={(e) => setUsername(e.target.value)}
+          />
+        </div>
+        <div>
+          <label>Email</label>
+          <input
+            type="email"
+            name="email"
+            placeholder="Enter Email"
+            value={email}
+            onChange={(e) => setEmail(e.target.value)}
+          />
+        </div>
+        <div>
+          <label>Password</label>
+          <input
+            type="password"
+            name="password"
+            placeholder="Enter Password"
+            value={password}
+            onChange={(e) => setPassword(e.target.value)}
+          />
+        </div>
+        <button type="submit">Submit</button>
+      </form>
+    </div>
+  );
+}
